--- lib/jxl/enc_jpeg_opt_data.cc
@@ -1,655 +1,0 @@
-// Copyright (c) the JPEG XL Project Authors. All rights reserved.
-//
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-#include "lib/jxl/enc_jpeg_opt_data.h"
-
-#include <cmath>
-
-#include "lib/jxl/ac_context.h"
-#include "lib/jxl/enc_Knuth_partition.h"
-
-namespace jxl {
-// Immutable data for JPEG recompression optimization.
-// Precomputed values of `n*log2(n)*kFScale` for `n` up to `max_n`.
-void JPEGOptData::InitFTab(size_t max_n) {
-  const size_t old_size = ftab.size();
-  if (max_n < old_size) return;
-  ftab.resize(max_n + 1, 0);
-  for (size_t i = std::max<size_t>(1, old_size); i <= max_n; ++i) {
-    double n = static_cast<double>(i);
-    ftab[i] = static_cast<int64_t>(std::llround(n * std::log2(n) * kFScale));
-  }
-}
-// `ftab` may not cover counts as large as `num_blocks`; fall back to
-// direct calculation for large values.
-int64_t JPEGOptData::NZFTab(uint32_t n) const {
-  if (n < ftab.size()) return ftab[n];
-  const double nd = static_cast<double>(n);
-  return static_cast<int64_t>(std::llround(nd * std::log2(nd) * kFScale));
-}
-
-// Map sparse `zdc_ai` to dense index `[0, num_zdcai)`.
-uint32_t JPEGOptData::CompactHBin(uint32_t zdc_ai) const {
-  JXL_DASSERT(zdc_ai < compact_map_h.size());
-  if (zdc_ai >= compact_map_h.size()) return kInvalidCompactH;
-  uint32_t dense = compact_map_h[zdc_ai];
-  JXL_DASSERT(dense != kInvalidCompactH);
-  return dense;
-}
-
-// Threshold initialization for one axis via the same contiguous-partition DP
-// used in later refinement, but with a cheaper 1D surrogate cost:
-// `cost[l..n] = f(sum_{i=l}^n DC_cnt[i])`.
-// The interval-cost matrix is written in the same diff form consumed by the
-// shared Knuth solver, so this stage reuses the exact same partition backend
-// as the AC-driven optimization path.
-Thresholds JPEGOptData::InitThresh(int axis, uint32_t target_intervals) const {
-  if (target_intervals <= 1) return {};
-  const Thresholds& dc_vals = DC_vals[axis];
-  const uint32_t M = static_cast<uint32_t>(dc_vals.size());
-  if (M == 0) return {};
-
-  // Fast exit: fewer distinct values than requested intervals.
-  if (M <= target_intervals) {
-    // Thresholds denote the first value of the next interval, so if every
-    // distinct DC value gets its own bucket we exclude the first one.
-    return Thresholds(dc_vals.begin() + 1, dc_vals.end());
-  }
-
-  std::vector<uint32_t> prefix(M + 1, 0);
-  for (uint32_t i = 0; i < M; ++i) {
-    prefix[i + 1] =
-        prefix[i] + DC_cnt[axis][static_cast<uint32_t>(dc_vals[i] + kDCTOff)];
-  }
-
-  KnuthPartitionSolver solver(M);
-  solver.ResetCosts(static_cast<size_t>(M) * M);
-  for (uint32_t n = 0; n < M; ++n) {
-    int64_t prev_base = 0;
-    for (uint32_t l = 0; l <= n; ++l) {
-      uint32_t total = prefix[n + 1] - prefix[l];
-      int64_t base = NZFTab(total);
-      if (n > 0 && l < n) {
-        base -= NZFTab(prefix[n] - prefix[l]);
-      }
-      solver.costs[n * M + l] = base - prev_base;
-      prev_base = base;
-    }
-  }
-
-  std::vector<uint32_t> split_points = solver.Solve(target_intervals, M);
-  Thresholds thresholds;
-  for (uint32_t split_point : split_points) {
-    thresholds.push_back(dc_vals[split_point]);
-  }
-  return thresholds;
-}
-
-// Find all factorizations `(a,b,c)` for DC thresholds.
-// Return only maximal factorizations (cannot increase any factor):
-// lower are not winning against higher factorizations for entropy.
-Factorizations JPEGOptData::MaximalFactorizations() const {
-  uint32_t cap0 = std::max(1u, static_cast<uint32_t>(DC_vals[0].size()));
-  uint32_t cap1 = std::max(1u, static_cast<uint32_t>(DC_vals[1].size()));
-  uint32_t cap2 = std::max(1u, static_cast<uint32_t>(DC_vals[2].size()));
-  Factorizations result;
-  for (uint32_t a = 1; a <= cap0 && a <= kMaxCells && a <= kMaxIntervals; ++a) {
-    for (uint32_t b = 1; b <= cap1 && a * b <= kMaxCells && b <= kMaxIntervals;
-         ++b) {
-      for (uint32_t c = 1;
-           c <= cap2 && a * b * c <= kMaxCells && c <= kMaxIntervals; ++c) {
-        bool can_inc_a =
-            (a < cap0) && (a < kMaxIntervals) && ((a + 1) * b * c <= kMaxCells);
-        bool can_inc_b =
-            (b < cap1) && (b < kMaxIntervals) && (a * (b + 1) * c <= kMaxCells);
-        bool can_inc_c =
-            (c < cap2) && (c < kMaxIntervals) && (a * b * (c + 1) <= kMaxCells);
-        if (!can_inc_a && !can_inc_b && !can_inc_c)
-          result.emplace_back(a, b, c);
-      }
-    }
-  }
-  // On greyscale images we need to join the rest two components into
-  // an additional empty context, so we have 1 less context for optimization,
-  // max 15 contexts
-  if (channels == 1) {
-    uint32_t ctxs = std::min(std::get<0>(result[0]), 15u);
-    std::get<0>(result[0]) = ctxs;
-    // while (--ctxs > 0) {
-    //   result.emplace_back(ctxs, 1, 1);
-    // }
-  }
-  return result;
-}
-
-// -------------------------------------------------------//
-// JPEG processing: each pass sweeps all image data once. //
-// -------------------------------------------------------//
-
-using ACCounts = JPEGOptData::ACCounts;
-// ------------------------------------------------------//
-// Pass 1: Count DC + AC values (parallel by component). //
-// ------------------------------------------------------//
-StatusOr<std::unique_ptr<ACCounts>> JPEGOptData::CountDCAC(
-    const jpeg::JPEGData& jpeg_data, ThreadPool* pool) {
-  auto ac_cnt = jxl::make_unique<ACCounts>();
-  memset(DC_cnt, 0, sizeof(DC_cnt));
-  memset(ac_cnt->data(), 0, sizeof(*ac_cnt));
-  JXL_RETURN_IF_ERROR(RunOnPool(
-      pool, 0, channels, ThreadPool::NoInit,
-      [&](uint32_t c, size_t) -> Status {
-        const auto& comp = jpeg_data.components[c];
-        uint32_t wb = comp.width_in_blocks;
-        uint32_t hb = comp.height_in_blocks;
-        for (uint32_t by = 0; by < hb; ++by) {
-          for (uint32_t bx = 0; bx < wb; ++bx) {
-            const int16_t* q = comp.coeffs.data() + (by * wb + bx) * 64;
-            ++DC_cnt[c][q[0] + kDCTOff];
-            for (uint8_t p = 0; p < kNumPos; ++p) {
-              ++(*ac_cnt)[c][p][q[jpeg::kJPEGNaturalOrder[p + 1]] + kDCTOff];
-            }
-          }
-        }
-        return true;
-      },
-      "CountDCAC"));
-
-  // Build DC value lists and LUTs.
-  for (auto& v : DC_vals) {
-    v.clear();
-    v.reserve(kDCTRange);
-  }
-  for (uint32_t c = 0; c < channels; ++c) {
-    auto& v = DC_vals[c];
-    for (int di = 0; di < kDCTRange; ++di) {
-      if (DC_cnt[c][di]) {
-        DC_idx_LUT[c][di] = static_cast<uint32_t>(v.size());
-        v.push_back(static_cast<int16_t>(di - kDCTOff));
-      }
-    }
-  }
-  return ac_cnt;
-}
-
-using ScanOrder = std::array<std::array<uint8_t, kNumPos>, kNumCh>;
-// JPEG XL high effort mode scans AC coefficients in descending nonzero
-// abundance (ties keep zigzag order).
-ScanOrder BuildACScanOrder(const ACCounts& ac_cnt) {
-  ScanOrder scan_order;
-  for (uint32_t c = 0; c < kNumCh; ++c) {
-    std::array<uint8_t, kNumPos> pos_order;
-    for (uint8_t i = 0; i < kNumPos; ++i) pos_order[i] = i;
-    std::stable_sort(pos_order.begin(), pos_order.end(),
-                     [&](uint8_t a, uint8_t b) {
-                       return ac_cnt[c][a][kDCTOff] < ac_cnt[c][b][kDCTOff];
-                     });
-    for (uint8_t s = 0; s < kNumPos; ++s) {
-      scan_order[c][s] = jpeg::kJPEGNaturalOrder[pos_order[s] + 1];
-    }
-  }
-  return scan_order;
-}
-
-// -------------------------------------------------------//
-// Pass 2: Collect AC bins, DC indices and `nz` per block //
-//         (parallel by comp).                            //
-// -------------------------------------------------------//
-Status JPEGOptData::BuildBlockOptData(const jpeg::JPEGData& jpeg_data,
-                                      ThreadPool* pool,
-                                      const ACCounts& ac_cnt) {
-  ScanOrder scan_order = BuildACScanOrder(ac_cnt);
-
-  JXL_RETURN_IF_ERROR(RunOnPool(
-      pool, 0, channels, ThreadPool::NoInit,
-      [&](uint32_t c, size_t) -> Status {
-        const auto& comp = jpeg_data.components[c];
-        uint32_t wb = comp.width_in_blocks;
-        uint32_t hb = comp.height_in_blocks;
-
-        block_DC_idx[c].assign(num_blocks[c], 0);
-        block_offsets[c].resize(num_blocks[c] + 1);
-        block_nonzeros[c].assign(num_blocks[c], 0);
-        block_nz_pred_bucket[c].resize(num_blocks[c]);
-        block_bins[c].clear();
-        std::vector<uint32_t> row_top(wb, 32u);
-        std::vector<uint32_t> row_cur(wb, 32u);
-        bool has_top = false;
-        uint32_t bc = 0;
-        for (uint32_t by = 0; by < hb; ++by) {
-          for (uint32_t bx = 0; bx < wb; ++bx, ++bc) {
-            block_offsets[c][bc] = static_cast<uint32_t>(block_bins[c].size());
-
-            const int16_t* q = comp.coeffs.data() + (by * wb + bx) * 64;
-            uint16_t DC_idx =
-                static_cast<uint16_t>(DC_idx_LUT[c][q[0] + kDCTOff]);
-            block_DC_idx[c][bc] = DC_idx;
-
-            uint32_t nonzeros_left = 0;
-            for (uint32_t s = 1; s <= kNumPos; ++s)
-              if (q[s] != 0) ++nonzeros_left;
-            block_nonzeros[c][bc] = static_cast<uint8_t>(nonzeros_left);
-
-            uint32_t predicted_nz;
-            if (bx == 0) {
-              predicted_nz = has_top ? row_top[bx] : 32u;
-            } else if (!has_top) {
-              predicted_nz = row_cur[bx - 1];
-            } else {
-              predicted_nz = (row_top[bx] + row_cur[bx - 1] + 1u) / 2u;
-            }
-            block_nz_pred_bucket[c][bc] = static_cast<uint8_t>(
-                (predicted_nz < 8) ? predicted_nz : (4 + predicted_nz / 2));
-            row_cur[bx] = nonzeros_left;
-
-            for (uint32_t s = 0; s < kNumPos; ++s) {
-              if (nonzeros_left == 0) break;
-              int16_t coeff = q[scan_order[c][s]];
-              bool nz_prev = (s > 0 && q[scan_order[c][s - 1]] != 0) ||
-                             (s == 0 && nonzeros_left > 4);
-              uint32_t zdc = static_cast<uint32_t>(
-                  ZeroDensityContext(nonzeros_left, s + 1, 1, 0, nz_prev));
-              uint32_t ai = static_cast<uint32_t>(coeff + kDCTOff);
-              block_bins[c].push_back((c << 20) | (zdc << 11) | ai);
-              nonzeros_left -= (coeff != 0);
-            }
-          }
-          row_top.swap(row_cur);
-          has_top = true;
-        }
-        block_offsets[c][num_blocks[c]] =
-            static_cast<uint32_t>(block_bins[c].size());
-        return true;
-      },
-      "CollectBlockOptData"));
-  return true;
-}
-
-// ------------------------------------------------------------------//
-// Pass 3: Prune DC values - remove DC values with no active bins.   //
-//         Build per-axis sorted block positions (parallel by comp). //
-// ------------------------------------------------------------------//
-Status JPEGOptData::FinalizeSpatialIndexing(ThreadPool* pool) {
-  memset(DC_cnt, 0, sizeof(DC_cnt));
-
-  // Loop over image blocks
-  if (channels == 1) {
-    for (uint32_t b = 0; b < num_blocks[0]; ++b) {
-      uint16_t dc0 = block_DC_idx[0][b];
-
-      ++DC_cnt[0][DC_vals[0][dc0] + kDCTOff];
-    }
-  } else {
-    for (uint32_t c = 0; c < kNumCh; ++c) {
-      for (uint32_t y = 0; y < block_grid_h[c]; ++y) {
-        for (uint32_t x = 0; x < block_grid_w[c]; ++x) {
-          uint32_t by = y * ss_y[c];
-          uint32_t bx = x * ss_x[c];
-          uint16_t dc0 =
-              block_DC_idx[0][(by / ss_y[0]) * block_grid_w[0] + bx / ss_x[0]];
-          uint16_t dc1 =
-              block_DC_idx[1][(by / ss_y[1]) * block_grid_w[1] + bx / ss_x[1]];
-          uint16_t dc2 =
-              block_DC_idx[2][(by / ss_y[2]) * block_grid_w[2] + bx / ss_x[2]];
-
-          ++DC_cnt[0][DC_vals[0][dc0] + kDCTOff];
-          ++DC_cnt[1][DC_vals[1][dc1] + kDCTOff];
-          ++DC_cnt[2][DC_vals[2][dc2] + kDCTOff];
-        }
-      }
-    }
-  }
-
-  JXL_RETURN_IF_ERROR(RunOnPool(
-      pool, 0, channels, ThreadPool::NoInit,
-      [&](uint32_t c, size_t) -> Status {
-        Thresholds pruned;
-        uint32_t idx = 0;
-        for (int di = 0; di < kDCTRange; ++di) {
-          if (DC_cnt[c][di]) {
-            pruned.push_back(static_cast<int16_t>(di - kDCTOff));
-            DC_idx_LUT[c][di] = idx++;
-          }
-        }
-        uint32_t M = static_cast<uint32_t>(pruned.size());
-        DC_block_offsets[c].assign(M + 1, 0);
-        for (uint16_t& dc_idx : block_DC_idx[c]) {
-          dc_idx = DC_idx_LUT[c][DC_vals[c][dc_idx] + kDCTOff];
-          ++DC_block_offsets[c][dc_idx + 1];
-        }
-        for (uint32_t v = 0; v < M; ++v)
-          DC_block_offsets[c][v + 1] += DC_block_offsets[c][v];
-        DC_vals[c].swap(pruned);
-        return true;
-      },
-      "PruneDC"));
-
-  JXL_RETURN_IF_ERROR(RunOnPool(
-      pool, 0, channels, ThreadPool::NoInit,
-      [&](uint32_t c, size_t) -> Status {
-        DC_sorted_blocks[c].resize(num_blocks[c]);
-        std::vector<uint32_t> write_pos = DC_block_offsets[c];
-        for (uint32_t y = 0; y < block_grid_h[c]; ++y) {
-          for (uint32_t x = 0; x < block_grid_w[c]; ++x) {
-            uint32_t b = y * block_grid_w[c] + x;
-            uint32_t v = block_DC_idx[c][b];
-            DC_sorted_blocks[c][write_pos[v]++] = (y << 16) | x;
-          }
-        }
-        return true;
-      },
-      "BuildSortedBlocks"));
-  return true;
-}
-
-// ---------------------------------------------------------------------//
-// Prepare AC stream via two-pass counting sort on the 22-bit bin key   //
-// `(c:2b, zdc:9b, ai:11b)` and emit on the third pass.                 //
-//                                                                      //
-// Pass 4: Find bin positions in the stream with pruned DC indices.     //
-// Pass 5: Scatter into `flat_lo` and `flat_hi` with pruned DC indices. //
-// Pass 6: Emit AC stream with run-length encoding.                     //
-// ---------------------------------------------------------------------//
-Status JPEGOptData::GenerateRLEStream(ThreadPool* pool) {
-  const uint32_t BIN_N = channels << 20;
-  // Split on `dc0` bit 10 (the MSB of the 11-bit index):
-  // `lo = dc0 < 1024`, `hi = dc0 >= 1024`.
-  // Each half packs `(dc0_low10 << 22 | dc1 << 11 | dc2)` into
-  // a single `uint32_t` (10+11+11 = 32 bits), halving `flat[]` memory
-  // vs the full `uint64_t` encoding.
-  std::vector<uint32_t> bin_start_lo(BIN_N + 1, 0);
-  std::vector<uint32_t> bin_start_hi(BIN_N + 1, 0);
-
-  for (uint32_t c = 0; c < channels; ++c) {
-    uint32_t w = block_grid_w[c];
-    uint32_t h = block_grid_h[c];
-    for (uint32_t y = 0; y < h; ++y) {
-      for (uint32_t x = 0; x < w; ++x) {
-        uint32_t b = y * w + x;
-        uint32_t b0 =
-            (y * ss_y[c] / ss_y[0]) * block_grid_w[0] + (x * ss_x[c] / ss_x[0]);
-        uint32_t dc0 = block_DC_idx[0][b0];
-        for (uint32_t pi = block_offsets[c][b]; pi < block_offsets[c][b + 1];
-             ++pi) {
-          uint32_t bin = block_bins[c][pi];
-          if (dc0 & 0x400u) {
-            ++bin_start_hi[bin + 1];
-          } else {
-            ++bin_start_lo[bin + 1];
-          }
-        }
-      }
-    }
-  }
-
-  // Prefix sums and active bins
-  std::vector<uint32_t> active_bins;
-  active_bins.reserve(BIN_N);
-  for (uint32_t b = 0; b < BIN_N; ++b) {
-    uint32_t c_lo = bin_start_lo[b + 1];
-    uint32_t c_hi = bin_start_hi[b + 1];
-    bin_start_lo[b + 1] = bin_start_lo[b] + c_lo;
-    bin_start_hi[b + 1] = bin_start_hi[b] + c_hi;
-    if (c_lo != 0 || c_hi != 0) active_bins.push_back(b);
-  }
-  active_bins.shrink_to_fit();
-
-  // `AC_stream` layout:
-  //   Regular frame (bit 31 = 0):
-  //     30..27  `Δdc0`   (4b, 0..15)
-  //     26..16  `dc1`    (11b, absolute)
-  //     15..5   `dc2`    (11b, absolute)
-  //     4..0    `run-1`  (5b; 0..30 = run 1..31; 31 = long-run escape)
-  //   Long-run frame (follows regular frame with `run-1 = 31`):
-  //     31..0   `run`    (32b, 2^26 ≥ actual run ≥ 32)
-  //   Reset frame (bit 31 = 1),
-  //     emitted for the first entry, bin change, `Δdc0 > 15`:
-  //     30      `ctx_changed` (1b)
-  //     29      `bin_changed` (1b)
-  //     28..7   `bin`      (22b = `c:2 + zdc:9 + ai:11`, used in clustering)
-  //     6..0    `dc0 >> 4` (7b, coarse; fine bits recovered from next `Δdc0`)
-  //
-  // The layout gives max 5% stream overhead in the worst case tested,
-  // dominated by the reset frames for `Δdc0 > 15`
-
-  // --------------------------------------------------------------------
-  // Pass 5: Scatter into `flat_lo` and `flat_hi` with pruned DC indices.
-  // --------------------------------------------------------------------
-  size_t lo_size = bin_start_lo[BIN_N];
-  size_t hi_size = bin_start_hi[BIN_N];
-  size_t raw_size = lo_size + hi_size;
-  std::vector<uint32_t> flat_lo(lo_size);
-  std::vector<uint32_t> flat_hi(hi_size);
-
-  // Scatter pass: `dc0 & 0x3FF` drops bit 10 (already encoded by lo/hi half);
-  // `dc0_base` (0 or 0x400) re-adds it during decode in `emit_half` later.
-  if (channels == 1) {
-    for (uint32_t b = 0; b < num_blocks[0]; ++b) {
-      uint32_t dc0 = block_DC_idx[0][b];
-      for (uint32_t pi = block_offsets[0][b]; pi < block_offsets[0][b + 1];
-           ++pi) {
-        uint32_t bin = block_bins[0][pi];
-        uint32_t dc_key = (dc0 & 0x3FFu) << 22;
-        if (dc0 & 0x400u) {
-          flat_hi[bin_start_hi[bin]++] = dc_key;
-        } else {
-          flat_lo[bin_start_lo[bin]++] = dc_key;
-        }
-      }
-    }
-  } else {
-    for (uint32_t c = 0; c < kNumCh; ++c) {
-      uint32_t w = block_grid_w[c];
-      uint32_t h = block_grid_h[c];
-      for (uint32_t y = 0; y < h; ++y) {
-        for (uint32_t x = 0; x < w; ++x) {
-          uint32_t b = y * w + x;
-          uint32_t b0 = (y * ss_y[c] / ss_y[0]) * block_grid_w[0] +
-                        (x * ss_x[c] / ss_x[0]);
-          uint32_t b1 = (y * ss_y[c] / ss_y[1]) * block_grid_w[1] +
-                        (x * ss_x[c] / ss_x[1]);
-          uint32_t b2 = (y * ss_y[c] / ss_y[2]) * block_grid_w[2] +
-                        (x * ss_x[c] / ss_x[2]);
-
-          uint32_t dc0 = block_DC_idx[0][b0];
-          uint32_t dc1 = block_DC_idx[1][b1];
-          uint32_t dc2 = block_DC_idx[2][b2];
-          for (uint32_t pi = block_offsets[c][b]; pi < block_offsets[c][b + 1];
-               ++pi) {
-            uint32_t bin = block_bins[c][pi];
-            uint32_t dc_key = ((dc0 & 0x3FFu) << 22) | (dc1 << 11) | dc2;
-            if (dc0 & 0x400u) {
-              flat_hi[bin_start_hi[bin]++] = dc_key;
-            } else {
-              flat_lo[bin_start_lo[bin]++] = dc_key;
-            }
-          }
-        }
-      }
-    }
-  }
-
-  // Sort each bin's lo/hi halves in parallel.
-  uint32_t kChunk = 1024;
-  JXL_RETURN_IF_ERROR(RunOnPool(
-      pool, 0,
-      static_cast<uint32_t>((active_bins.size() + kChunk - 1) / kChunk),
-      ThreadPool::NoInit,
-      [&](uint32_t chunk, size_t) -> Status {
-        uint32_t idx_start = chunk * kChunk;
-        uint32_t idx_end = std::min(idx_start + kChunk,
-                                    static_cast<uint32_t>(active_bins.size()));
-        for (uint32_t idx = idx_start; idx < idx_end; ++idx) {
-          uint32_t b = active_bins[idx];
-          uint32_t s_lo = (b == 0) ? 0 : bin_start_lo[b - 1];
-          uint32_t e_lo = bin_start_lo[b];
-          uint32_t s_hi = (b == 0) ? 0 : bin_start_hi[b - 1];
-          uint32_t e_hi = bin_start_hi[b];
-          if (s_lo < e_lo)
-            std::sort(flat_lo.data() + s_lo, flat_lo.data() + e_lo);
-          if (s_hi < e_hi)
-            std::sort(flat_hi.data() + s_hi, flat_hi.data() + e_hi);
-        }
-        return true;
-      },
-      "SortBins"));
-
-  // ------------------------------------------------
-  // Pass 6: Emit AC stream with run-length encoding.
-  // ------------------------------------------------
-  std::vector<ACEntry> stream;
-  stream.reserve(raw_size + raw_size / 16);
-  uint32_t ctx_len = 0;
-  uint32_t prev_bin = UINT32_MAX;
-  std::array<uint32_t, kZeroDensityContextCount> zdc_total = {};
-
-  // `compact_map_h` maps sparse `(zdc<<11|ai)` keys (20-bit, max ~1M) to a
-  // dense id `[0, num_unique_h)`. Only keys that appear in `active_bins`
-  // get assigned; the rest remain 0xFFFFFFFF (unused sentinel).
-  // Actual number of active bins seen is ~35000.
-  const uint32_t max_key_h_sparse =
-      static_cast<uint32_t>(kZeroDensityContextCount * kDCTRange);
-  compact_map_h.assign(max_key_h_sparse, kInvalidCompactH);
-  dense_to_zdcai.clear();
-  dense_to_zdcai.reserve(active_bins.size());
-  num_zdcai = 0;
-
-  uint32_t lo_pos = 0;
-  uint32_t hi_pos = 0;
-  for (unsigned int bin : active_bins) {
-    uint32_t zdc_ai =
-        bin & 0xFFFFFu;  // strip channel bits (top 2 bits of `bin`)
-    if (compact_map_h[zdc_ai] == kInvalidCompactH) {
-      compact_map_h[zdc_ai] = num_zdcai++;
-      dense_to_zdcai.push_back(zdc_ai);
-    }
-
-    uint32_t start_lo = lo_pos;
-    uint32_t start_hi = hi_pos;
-    uint32_t end_lo = bin_start_lo[bin];
-    uint32_t end_hi = bin_start_hi[bin];
-    lo_pos = end_lo;
-    hi_pos = end_hi;
-
-    // Stats: track per-ctx entry counts.
-    // `bin >> 11` extracts `(c<<9|zdc)`, which identifies the context.
-    bool new_ctx =
-        (prev_bin == UINT32_MAX) || ((bin >> 11) != (prev_bin >> 11));
-    if (new_ctx) {
-      if (prev_bin != UINT32_MAX) {
-        zdc_total[(prev_bin >> 11) & 0x1FFu] += ctx_len;
-      }
-      ctx_len = 0;
-    }
-    ctx_len += (end_lo - start_lo) + (end_hi - start_hi);
-
-    // `bin_change` / `ctx_change` mark reset-frame flags (bits 29/30).
-    // `bin_change` is false only for the very first bin in the entire stream.
-    bool bin_change = (prev_bin != UINT32_MAX);
-    bool ctx_change = bin_change && ((bin >> 11) != (prev_bin >> 11));
-
-    // Emit one sorted half (lo or hi) into stream.
-    // `dc0_base` is added back to recover the full 11-bit `dc0`.
-    uint32_t cur_dc0 = 0;
-    bool first_in_bin = true;
-    auto emit_half = [&](const std::vector<uint32_t>& data, uint32_t start,
-                         uint32_t end, uint32_t dc0_base) {
-      uint32_t i = start;
-      while (i < end) {
-        uint32_t j = i + 1;
-        while (j < end && data[j] == data[i]) ++j;
-        uint32_t run = j - i;
-        uint32_t dc0 = (data[i] >> 22) | dc0_base;
-        uint32_t dc1 = (data[i] >> 11) & 0x7FFu;
-        uint32_t dc2 = data[i] & 0x7FFu;
-
-        // Reset frame: emitted on first entry, bin change, or `Δdc0 > 15`.
-        // `cur_dc0` is snapped to the coarsest 16-aligned value ≤ dc0
-        // so the 4-bit `Δdc0` in the next normal frame is always in [0,15].
-        if (first_in_bin || (dc0 - cur_dc0 > 15u)) {
-          stream.push_back(
-              (1u << 31) |
-              (static_cast<uint32_t>(ctx_change && first_in_bin) << 30) |
-              (static_cast<uint32_t>(bin_change && first_in_bin) << 29) |
-              (bin << 7) | (dc0 >> 4));
-          cur_dc0 = (dc0 >> 4) << 4;
-        }
-
-        // Normal frame: `Δdc0` in bits 30..27, `dc1` in bits 26..16,
-        // `dc2` in bits 15..5.
-        uint32_t delta_dc0 = dc0 - cur_dc0;
-        uint32_t header = (delta_dc0 << 27) | (dc1 << 16) | (dc2 << 5);
-        if (run <= 31) {
-          stream.push_back(header | (run - 1));  // `run-1` fits in 5 bits
-        } else {
-          stream.push_back(header | 0x1Fu);  // escape: `run-1 == 31`
-          stream.push_back(run);  // long-run frame carries actual run
-        }
-        cur_dc0 = dc0;
-        first_in_bin = false;
-        i = j;
-      }
-    };
-
-    emit_half(flat_lo, start_lo, end_lo, 0u);
-    emit_half(flat_hi, start_hi, end_hi, 0x400u);
-    prev_bin = bin;
-  }
-
-  if (prev_bin != UINT32_MAX) {
-    zdc_total[(prev_bin >> 11) & 0x1FFu] += ctx_len;
-  }
-  stream.shrink_to_fit();
-  AC_stream = std::move(stream);
-
-  // Clustering may merge contexts from different channels, so
-  // `hist_N[merged][zdc]` can accumulate counts from all channels for the
-  // same `zdc`. Use the actual per-`zdc` cross-channel total as the bound
-  // for `ftab` size.
-  uint32_t max_zdc_total =
-      *std::max_element(zdc_total.begin(), zdc_total.end());
-  InitFTab(max_zdc_total + 1);
-
-  return true;
-}
-
-// Orchestrates the JPEG data extraction pipeline.
-Status JPEGOptData::BuildFromJPEG(const jpeg::JPEGData& jpeg_data,
-                                  ThreadPool* pool) {
-  channels = static_cast<uint32_t>(jpeg_data.components.size());
-  w_max = 0;
-  h_max = 0;
-  for (uint32_t c = 0; c < channels; ++c) {
-    block_grid_w[c] = jpeg_data.components[c].width_in_blocks;
-    block_grid_h[c] = jpeg_data.components[c].height_in_blocks;
-    num_blocks[c] = block_grid_w[c] * block_grid_h[c];
-    w_max = std::max(w_max, block_grid_w[c]);
-    h_max = std::max(h_max, block_grid_h[c]);
-  }
-  for (uint32_t c = 0; c < channels; ++c) {
-    ss_y[c] = h_max / block_grid_h[c];
-    ss_x[c] = w_max / block_grid_w[c];
-  }
-  for (uint32_t c = channels; c < kNumCh; ++c) {
-    block_grid_w[c] = 0;
-    block_grid_h[c] = 0;
-    num_blocks[c] = 0;
-    ss_y[c] = 0;
-    ss_x[c] = 0;
-  }
-
-  {
-    JXL_ASSIGN_OR_RETURN(auto ac_cnt, CountDCAC(jpeg_data, pool));
-
-    JXL_RETURN_IF_ERROR(BuildBlockOptData(jpeg_data, pool, *ac_cnt));
-  }
-
-  JXL_RETURN_IF_ERROR(FinalizeSpatialIndexing(pool));
-
-  JXL_RETURN_IF_ERROR(GenerateRLEStream(pool));
-
-  return true;
-}
-
-}  // namespace jxl
